Fixed the fluid multimeter recipe. Added ability to enable / disable the sharing of energy using EnergyUtils.giveEnergyAllFaces and takeEnergyAllFaces. Made sure that Tesla has priority over CoFH and Forge Energy to allow for 64 bit energy storage. Added russian language.

+++ src/main/java/cjminecraft/core/energy/compat/TileEntityEnergyProducer.java
@@ -32,7 +32,7 @@ public class TileEntityEnergyProducer extends TileEntityEnergy implements IEnerg
 	 * @param capacity
 	 *            The capacity of the energy storage
 	 */
-	public TileEntityEnergyProducer(int capacity) {
+	public TileEntityEnergyProducer(long capacity) {
 		super(capacity, capacity, capacity, 0);
 	}
 
@@ -44,7 +44,7 @@ public class TileEntityEnergyProducer extends TileEntityEnergy implements IEnerg
 	 * @param maxTransfer
 	 *            The max receive and max extract of the energy storage
 	 */
-	public TileEntityEnergyProducer(int capacity, int maxTransfer) {
+	public TileEntityEnergyProducer(long capacity, long maxTransfer) {
 		super(capacity, maxTransfer, maxTransfer, 0);
 	}
 
@@ -58,7 +58,7 @@ public class TileEntityEnergyProducer extends TileEntityEnergy implements IEnerg
 	 * @param maxExtract
 	 *            The maximum amount of energy which can be extracted
 	 */
-	public TileEntityEnergyProducer(int capacity, int maxReceive, int maxExtract) {
+	public TileEntityEnergyProducer(long capacity, long maxReceive, long maxExtract) {
 		super(capacity, maxReceive, maxExtract, 0);
 	}
 
@@ -74,20 +74,20 @@ public class TileEntityEnergyProducer extends TileEntityEnergy implements IEnerg
 	 * @param energy
 	 *            The energy inside of the energy storage
 	 */
-	public TileEntityEnergyProducer(int capacity, int maxReceive, int maxExtract, int energy) {
+	public TileEntityEnergyProducer(long capacity, long maxReceive, long maxExtract, long energy) {
 		super(capacity, maxReceive, maxExtract, energy);
 	}
 
 	@Optional.Method(modid = "redstoneflux")
 	@Override
 	public int getEnergyStored(EnumFacing from) {
-		return this.storage.getEnergyStored();
+		return (int) this.storage.getEnergyStored();
 	}
 
 	@Optional.Method(modid = "redstoneflux")
 	@Override
 	public int getMaxEnergyStored(EnumFacing from) {
-		return this.storage.getMaxEnergyStored();
+		return (int) this.storage.getMaxEnergyStored();
 	}
 
 	@Optional.Method(modid = "redstoneflux")
@@ -99,7 +99,7 @@ public class TileEntityEnergyProducer extends TileEntityEnergy implements IEnerg
 	@Optional.Method(modid = "redstoneflux")
 	@Override
 	public int extractEnergy(EnumFacing from, int maxExtract, boolean simulate) {
-		return this.storage.extractEnergy(maxExtract, simulate);
+		return (int) this.storage.extractEnergy(maxExtract, simulate);
 	}
 
 	@Override
